Copy directly from cache to destination file instead of via tmp file
ccache copies files to the cache via a temporary file which is then
renamed into place. For simplicity reasons the same file copy function
is used when copying files from the cache as well. This however makes
ccache fail for very long destination filenames.

Fix this by introducing a “via_tmp_file” parameter to copy_file, similar
to how it works on master.

Closes #593.