Refactor result code into a Result::Reader/Writer classes
Introduced a Result::Reader class which remove knowledge about what to
do with the individual file parts of a result. That’s instead defined by
a Result::Reader::Consumer instance passed to Result::Reader::read. This
means a ResultFileMap is no longer used and that the consumer is free to
write data whereever it wants.

The new ResultRetriever class implements retrieval by writing to local
files like before, except that the stderr data is written directly to
STDERR_FILENO instead of landing in a temporary file.

The new ResultDumper class implements “ccache --dump-result”.

Also introduced a Result::Writer class, mostly for symmetry with the
Result::Reader class.