Remove unnecessary compiler behavior test in modules test suite
The “preprocessor output” test in the modules test suite only tests
compiler behavior, and the string it looks for in the preprocessed
output is different for older and newer Clang versions. It seems like a
kind of a probing call, but I don’t see why we would want to probe the
compiler behavior in this case so I’ll just remove it.

Fixes #601.