Improve handling of ccache version in the source release archive
The version.cpp file is currently generated in the source directory.
This is a bit unclean since generated files are supposed to be put in
the build directory.

Also, when building from a source release archive outside a Git
repository the CMake scripts still try to get the version from Git and
then emit an ugly warning message:

    fatal: not a git repository (or any of the parent directories): .git
    fatal: not a git repository (or any of the parent directories): .git
    CMake Warning at cmake/GenerateVersionFile.cmake:42 (message):
      Running git failed
    Call Stack (most recent call first):
      cmake/GenerateVersionFile.cmake:50 (get_version_from_git)
      CMakeLists.txt:32 (include)

Fix this by including a VERSION file in the source release archive. If
the VERSION file exists, GenerateVersionFile.cmake just uses the version
from the file and doesn’t try to get a version from Git. If the file
doesn’t exist, the version is looked up from Git like before, but it’s a
fatal error if the version cannot be determined.