feat: Make --config-path and --directory affect whole command line
Command line options are processed strictly in order, which means that

    ccache -d dir -s

is not the same as

    ccache -s -d dir

which is not very intuitive. There are also planned features where is
doesn’t make sense to have to put “non-command options” before the
“command options”.

Improve this by processing command line options in two passes: first
non-command options and then command options (still in order).