test: Use shell builtin pwd for basedir test
`/bin/pwd` won't exist on some systems (eg. NixOS). The only
difference between the bash builtin and the binary is that the binary
uses `-P` by default. By passing `-P` we can be sure that we get the
actual path, and not the symlinked path, which was needed to fix #916.