fix: normalize abs paths when collecting includes to fix basedir option
On windows with basedir to set e.g. C:\Projects, it will
get normalized to C:/Projects. With a header in C:\Projects\repo\some_header.h,
during include reading it may be read as C:\\\\Projects\\\\\repo\\\\some_header.h
with (extra) backslashes. It does not then match basedir which contains forward
slashes and path does not changed to relative when it should be.

Also the header paths printed to stdout need to be changed to relative so that Ninja will
use the relative paths to headers to detect changes instead of absolute paths. Otherwise
Ninja will check for header changes from the original build dir if an object was restored
from cache to another build dir.