chore: Remove share-hits attribute for secondary storage
[1] added a share-hits attribute for secondary storages so that it's
possible to avoid sharing hits to primary storage for a specific
secondary storage. I believe that nobody needs that level of control --
what one would like is the ability to not use the primary storage at
all. Such a feature will be added in a a future commit, but for now the
share-hits=false functionality is just in the way, so let's remove it.

[1]: 1924ad69cc42336a61ee69d0042de53db2a2a52c

(cherry picked from commit fd91ae11db7b49ad26a347f08517b81c7d549153)