feat: Add basic support for yacc/bison files
Bison takes as input a .y file and produces a .c file, and optionally
also a .h file if "-d" is provided.  The .y file does have CPP #include
directives in regions of C code, but those survive unexpanded in the
output .c file (for a later build stage to deal with), so "direct mode"
seems to be sufficient.

This has been tested with PostgreSQL using BISON="ccache bison".

XXX Work in progress!